Create API generation abstract base class
